refactor: merge web2helper interceptor with messenger interceptor
