+++ src/contentscript/document-start/messenger.interceptor.ts
@@ -1,13 +1,55 @@
-import { InpageReqMessageFormat, ResponseMessageFormat, ResponseWithMessage } from '../../utils/message/message-handler'
+import {
+  InpageReqMessageFormat,
+  InterceptorReqMessageFormat,
+  InterceptorResMessageFormat,
+  ResponseMessageFormat,
+  ResponseWithMessage,
+} from '../../utils/message/message-handler'
 
 export class MessengerInterceptor {
-  protected readonly inpageOrigin = window.location.origin
+  private readonly inpageOrigin = window.location.origin
 
-  protected validInpageMessage(message: MessageEvent<InpageReqMessageFormat<unknown>>): boolean {
+  constructor() {
+    this.serveEvents()
+  }
+  serveEvents(): void {
+    console.log('Register Web2HelperInterceptor event listeners...')
+
+    window.addEventListener('message', (message: MessageEvent<InpageReqMessageFormat<undefined>>) => {
+      if (!this.validInpageMessage(message)) return
+
+      console.log('Web2HelperInterceptor content script got aimed message', message.data)
+      const messageToBackground: InterceptorReqMessageFormat = {
+        key: message.data.key,
+        sender: 'content',
+        target: 'background',
+      }
+      const messageToInpage: InterceptorResMessageFormat<string> = {
+        key: message.data.key,
+        eventId: message.data.eventId,
+        sender: 'content',
+        target: 'inpage',
+      }
+      chrome.runtime.sendMessage(messageToBackground, response => {
+        try {
+          const responseMessage = this.deserializeResponseMessage<string>(response)
+
+          messageToInpage.answer = responseMessage.answer
+          window.postMessage(messageToInpage, this.inpageOrigin)
+        } catch (e) {
+          messageToInpage.error = e.message
+          console.error('Web2HelperInterceptor Error', messageToInpage.error)
+          window.postMessage(messageToInpage, this.inpageOrigin)
+        }
+      })
+    })
+  }
+
+  private validInpageMessage(message: MessageEvent<InpageReqMessageFormat<unknown>>): boolean {
     if (
-      message.data.eventId ||
-      message.data.key ||
-      message.data.target === 'content' ||
+      message.data.eventId &&
+      message.data.key &&
+      message.data.target === 'content' &&
       message.data.sender === 'inpage'
     ) {
       return true
@@ -16,7 +58,7 @@ export class MessengerInterceptor {
     return false
   }
 
-  protected deserializeResponseMessage<T>(message: ResponseMessageFormat<T>): ResponseWithMessage<T> {
+  private deserializeResponseMessage<T>(message: ResponseMessageFormat<T>): ResponseWithMessage<T> {
     if (message.error) {
       throw new Error(message.error)
     }
